Tidied up in_process_param_store.py. Migration code move to migrations.py
