+++ entropylab/pipeline/results_backend/sqlalchemy/alembic/versions/2022-05-19-09-12-01_06140c96c8c4_wrapping_param_store_values.py
@@ -11,9 +11,9 @@ import shutil
 from entropylab.logger import logger
 from entropylab.pipeline.api.errors import EntropyError
 from entropylab.pipeline.api.in_process_param_store import (
-    migrate_param_store_0_1_to_0_2,
     _set_version,
 )
+from entropylab.pipeline.api.migrations import migrate_param_store_0_1_to_0_2
 from entropylab.pipeline.results_backend.sqlalchemy.alembic.alembic_util import (
     AlembicUtil,
 )
