+++ entropylab/pipeline/results_backend/sqlalchemy/alembic/versions/2022-06-28-12-13-39_09f3b5a1689c_fixing_param_qualified_name.py
@@ -10,9 +10,9 @@ import shutil
 from entropylab.logger import logger
 from entropylab.pipeline.api.errors import EntropyError
 from entropylab.pipeline.api.in_process_param_store import (
-    fix_param_qualified_name,
     _set_version,
 )
+from entropylab.pipeline.api.migrations import fix_param_qualified_name
 from entropylab.pipeline.results_backend.sqlalchemy.alembic.alembic_util import (
     AlembicUtil,
 )
